deriveErrorsAndWarnings
